Add AbstractBrooklynMojo
