+++ src/main/java/io/brooklyn/maven/mojo/AbstractInvokeBrooklynMojo.java
@@ -39,10 +39,7 @@ import com.google.common.annotations.VisibleForTesting;
 /**
  * An abstract class for Mojos that invoke actions on an existing Brooklyn server.
  */
-public abstract class AbstractInvokeBrooklynMojo extends AbstractMojo {
-
-    @Parameter(defaultValue = "${project}", readonly = true)
-    private MavenProject project;
+public abstract class AbstractInvokeBrooklynMojo extends AbstractBrooklynMojo {
 
     /**
      * The URL of the Brooklyn server to communicate with.
@@ -63,22 +60,6 @@ public abstract class AbstractInvokeBrooklynMojo extends AbstractMojo {
      */
     @Parameter(property = "brooklyn.password")
     protected String password;
-
-    /**
-     * The duration mojos should wait for actions at Brooklyn to complete.
-     */
-    @Parameter(
-            property = "brooklyn.timeout",
-            defaultValue = "5")
-    private Integer timeout;
-
-    /**
-     * The unit associated with {@link #timeout}.
-     */
-    @Parameter(
-            property = "brooklyn.timeoutUnit",
-            defaultValue = "MINUTES")
-    private TimeUnit timeoutUnit;
 
     /**
      * The period that should be waited between successive polls of the Brooklyn server.
@@ -106,10 +87,9 @@ public abstract class AbstractInvokeBrooklynMojo extends AbstractMojo {
     }
 
     public AbstractInvokeBrooklynMojo(URL server) {
+        super();
         this.server = server;
         // Values are set here to aid tests. They are overwritten when Maven invokes the plugin.
-        this.timeout = 5;
-        this.timeoutUnit = TimeUnit.SECONDS;
         this.pollPeriod = 1;
         this.pollUnit = TimeUnit.SECONDS;
     }
@@ -133,29 +113,10 @@ public abstract class AbstractInvokeBrooklynMojo extends AbstractMojo {
         return Duration.of(pollPeriod, pollUnit);
     }
 
-    protected MavenProject getProject() {
-        return project;
-    }
-
-    protected Duration getTimeout() {
-        return Duration.of(timeout, timeoutUnit);
-    }
 
     AbstractInvokeBrooklynMojo setPollPeriod(int period, TimeUnit unit) {
         this.pollPeriod = checkNotNull(period, "period");
         this.pollUnit = checkNotNull(unit, "unit");
-        return this;
-    }
-
-    @VisibleForTesting
-    AbstractInvokeBrooklynMojo setProject(MavenProject project) {
-        this.project = project;
-        return this;
-    }
-
-    AbstractInvokeBrooklynMojo setTimeout(int timeout, TimeUnit unit) {
-        this.timeout = checkNotNull(timeout, "timeout");
-        this.timeoutUnit = checkNotNull(unit, "unit");
         return this;
     }
